+++ test/battle/ability/aerilate.c
@@ -46,7 +46,7 @@ SINGLE_BATTLE_TEST("Aerilate can not turn certain moves into Flying type moves")
     }
 }
 
-SINGLE_BATTLE_TEST("Aerilate boosts power of affected moves by 20% (Gen7+) or 30% (Gen1-6)", s16 damage)
+SINGLE_BATTLE_TEST("Aerilate boosts power of affected moves by 20% (Gen7+) or 30% (Gen1-6)", s32 damage)
 {
     enum Move move;
     u32 genConfig;
@@ -73,7 +73,7 @@ SINGLE_BATTLE_TEST("Aerilate boosts power of affected moves by 20% (Gen7+) or 30
     }
 }
 
-SINGLE_BATTLE_TEST("Aerilate doesn't affect Weather Ball's type", s16 damage)
+SINGLE_BATTLE_TEST("Aerilate doesn't affect Weather Ball's type", s32 damage)
 {
     enum Move move1, move2;
     PARAMETRIZE { move1 = MOVE_CELEBRATE; move2 = MOVE_CELEBRATE; }
