+++ test/battle/ability/fur_coat.c
@@ -6,7 +6,7 @@ ASSUMPTIONS
     ASSUME(GetMoveCategory(MOVE_SCRATCH) == DAMAGE_CATEGORY_PHYSICAL);
 }
 
-SINGLE_BATTLE_TEST("Fur Coat doubles Defense", s16 damage)
+SINGLE_BATTLE_TEST("Fur Coat doubles Defense", s32 damage)
 {
     enum Ability ability;
     PARAMETRIZE { ability = ABILITY_FUR_COAT; }
@@ -24,7 +24,7 @@ SINGLE_BATTLE_TEST("Fur Coat doubles Defense", s16 damage)
     }
 }
 
-SINGLE_BATTLE_TEST("Fur Coat has no effect on self-inflicted confusion damage", s16 damage)
+SINGLE_BATTLE_TEST("Fur Coat has no effect on self-inflicted confusion damage", s32 damage)
 {
     enum Ability ability;
     PARAMETRIZE { ability = ABILITY_FUR_COAT; }
