+++ test/battle/ability/galvanize.c
@@ -43,7 +43,7 @@ SINGLE_BATTLE_TEST("Galvanize can not turn certain moves into Electric type move
     }
 }
 
-SINGLE_BATTLE_TEST("Galvanize boosts power of affected moves by 20% (Gen7+) or 30% (Gen1-6)", s16 damage)
+SINGLE_BATTLE_TEST("Galvanize boosts power of affected moves by 20% (Gen7+) or 30% (Gen1-6)", s32 damage)
 {
     enum Ability ability;
     u32 genConfig;
@@ -68,7 +68,7 @@ SINGLE_BATTLE_TEST("Galvanize boosts power of affected moves by 20% (Gen7+) or 3
     }
 }
 
-SINGLE_BATTLE_TEST("Galvanize doesn't affect Weather Ball's type", s16 damage)
+SINGLE_BATTLE_TEST("Galvanize doesn't affect Weather Ball's type", s32 damage)
 {
     enum Move move;
     enum Ability ability;
